adding docs and separate stories
